requirements2.txt
websockets
torch==1.12.1+cu113 
torchvision==0.13.1+cu113 
torchaudio==0.12.1 
opencv-python
onnx
onnxruntime-gpu
--extra-index-url https://download.pytorch.org/whl/cu113